:truck: Moved ConfigStruct.hpp in src/utils folder
