+++ plugins/systems/config/Config.hpp
@@ -8,7 +8,7 @@
 #ifndef CONFIGLOADER_HPP
     #define CONFIGLOADER_HPP
 
-    #include "ECS/config/ConfigStruct.hpp"
+    #include "utils/ConfigStruct.hpp"
     #include "plugins/systems/ASystem.hpp"
     #include <string>
     #include <libconfig.h++>
